+++ src/common/fire-event.ts
@@ -30,6 +30,7 @@
 
 declare global {
 	// tslint:disable-next-line
+	// eslint-disable-next-line @typescript-eslint/no-empty-object-type
 	interface HASSDomEvents {}
 }
