+++ src/lib/common.html.ts
@@ -1,6 +1,7 @@
 import { mdiCalendar } from '@mdi/js';
 import dayjs from 'dayjs';
 import { html } from 'lit';
+import { unsafeHTML } from 'lit/directives/unsafe-html.js';
 
 import EventClass from './event.class';
 import { getEventIcon } from '../helpers/get-icon';
@@ -150,5 +151,121 @@ export function getTitleHTML(config: atomicCardConfig, event: EventClass, hass: 
 				</div>
 			</a>
 		`;
+	}
+}
+
+/**
+ * generate Event Location link HTML
+ * @param config card configuration
+ * @param event event to get location from
+ * @returns TemplateResult containing location information
+ */
+export function getLocationHTML(config: atomicCardConfig, event: EventClass) {
+	if (!event.location || !config.showLocation) {
+		return html``;
+	} else if (config.disableLocationLink) {
+		return html`<ha-icon
+				class="event-location-icon"
+				style="--location-icon-color: ${config.locationIconColor}"
+				icon="mdi:map-marker"
+			></ha-icon
+			>&nbsp;${event.address}`;
+	} else {
+		const loc: string = event.location;
+		const location: string = loc.startsWith('http') ? loc : 'https://maps.google.com/?q=' + loc;
+		return html`<a
+			href=${location}
+			target="${config.linkTarget}"
+			class="location-link"
+			style="--location-link-size: ${config.locationTextSize}%"
+		>
+			<ha-icon
+				class="event-location-icon"
+				style="--location-icon-color: ${config.locationIconColor}"
+				icon="mdi:map-marker"
+			>
+			</ha-icon
+			>&nbsp;${event.address}
+		</a>`;
+	}
+}
+
+/**
+ * Gets the location for the event and displays if required
+ * @param config card configuration
+ * @param event event to get location for
+ * @returns TemplateResult with location icon and link
+ */
+export function getCalendarLocationHTML(config: atomicCardConfig, event: EventClass) {
+	if (!event.location || !config.showLocation || config.disableCalLocationLink) {
+		return html``;
+	} else {
+		const loc: string = event.location;
+		const location: string = loc.startsWith('http') ? loc : 'https://maps.google.com/?q=' + loc;
+		return html`
+			<a
+				href=${location}
+				target="${config.linkTarget}"
+				class="location-link"
+				style="--location-link-size: ${config.locationTextSize}%"
+			>
+				<ha-icon
+					class="event-location-icon"
+					style="--location-icon-color: ${config.locationIconColor}"
+					icon="mdi:map-marker"
+				></ha-icon
+				>&nbsp;
+			</a>
+		`;
+	}
+}
+
+export function getDescription(config: atomicCardConfig, event: EventClass) {
+	if (config.showDescription && event.description) {
+		let { description } = event;
+		if (isHtml(event.description)) {
+			description = unsafeHTML(event.description);
+		}
+
+		if (!isHtml(event.description) && config.descLength && event.description.length >= config.descLength) {
+			description = html`${event.description.slice(0, config.descLength)}`;
+		}
+		return html`<div class="event-right">
+			<div class="event-main">
+				<div
+					class="event-description"
+					style="--description-color: ${config.descColor}; --description-size: ${config.descSize}%"
+				>
+					${description}
+				</div>
+			</div>
+		</div>`;
+	}
+	return html``;
+}
+
+/**
+ * Gets the calendar description and wraps in html
+ * @param config calendar card configuration
+ * @param event event record
+ * @returns html description
+ */
+export function getCalendarDescriptionHTML(config: atomicCardConfig, event: EventClass) {
+	if (event.description) {
+		let { description } = event;
+		if (isHtml(event.description)) {
+			description = unsafeHTML(event.description);
+		}
+		if (!isHtml(event.description) && config.descLength && event.description.length > config.descLength) {
+			description = event.description.slice(0, config.descLength);
+		}
+		return html`<div
+			class="calDescription"
+			style="--description-color: ${config.descColor}; --description-size: ${config.descSize}%"
+		>
+			- ${description}
+		</div>`;
+	} else {
+		return html`;`;
 	}
 }
