+++ src/lib/loader-host.interface.ts
@@ -1,0 +1,4 @@
+export interface ILoaderHost {
+	showLoader: boolean;
+	requestUpdate(): void;
+}
